feat: OpenAPIから自動生成されるリクエストモデルにGoの型指定を定義できるようにする

+++ post-app/interface/controller/get_post.go
@@ -7,7 +7,7 @@ import (
 
 // GetPost 投稿詳細情報の取得
 // (GET /posts/{postId})
-func (s *Server) GetPost(c *gin.Context, postId int) {
+func (s *Server) GetPost(c *gin.Context, postId uint64) {
 	postInteractor := s.InteractProvider.ProvidePostInteractor(c)
 	presenter := presenter.NewPostPresenter(c)
 	postInteractor.GetPost(c, postId, presenter)
